+++ wemake_python_styleguide/visitors/ast/function_empty_lines.py
@@ -47,19 +47,20 @@ class _FileFunctions(object):
     def _search_functions(self) -> Iterator[_Function]:
         function_tokens: List[tokenize.TokenInfo] = []
         in_function = False
-        function_start_column = 0
+        function_start_token = (0, 0)
         for token in self._file_tokens:
             function_ended = self._is_function_end(
                 token,
                 bool(function_tokens),
-                function_start_column,
+                function_start_token[1],
+                function_start_token[0],
             )
             if not in_function and self._is_function_start(token):
                 in_function = True
-                function_start_column = token.start[1]
+                function_start_token = token.start
             elif function_ended:
                 in_function = False
-                function_start_column = 0
+                function_start_token = (function_start_token[0], 0)
                 yield _Function(function_tokens)
                 function_tokens = []
             if in_function:
@@ -73,7 +74,12 @@ class _FileFunctions(object):
         token: tokenize.TokenInfo,
         function_tokens_exists: bool,
         function_start_column: int,
+        function_start_line: int,
     ) -> bool:
+        is_elipsis = token.string == '...'
+        is_elipsis_end = is_elipsis and token.start[0] == function_start_line
+        if is_elipsis_end:
+            return True
         column_valid = token.start[1] in {0, function_start_column}
         is_dedent_token = token.type == tokenize.DEDENT
         return is_dedent_token and function_tokens_exists and column_valid