+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Semantic versioning in our case means:
 ### Bugfixes
 
 - Fix `ForbiddenInlineIgnoreViolation` config parsing. #2590
+- Fix `WrongEmptyLinesCountViolation` for func definitions with ellipsis. #2847
 
 
 ## 0.18.0